Switch test runner image to amd64 (#25)
This changes the test runner image to be based on a amd64 ubuntu:24.04
base image. To build students' solutions, we use a cross toolchain. To
run solutions, we use qemu-user. Students are asked to use the exact
same workflow. To this end we unify the tests Makefiles with those from
the arm64-assembly exercises repository.